Enable support for multi-ontology queries in query action #12 (#21)
* Refactor query filtering with expression groups

Introduces Expr and ExprGroup abstractions for query filtering, replacing direct use of OntologyFilter in query, repo, and timeseries modules.
Updates SQL query builders and compilers to use the new types, and refactors the query facade and endpoint to leverage FacadeQuery for unified query logic.
Also improves type safety and code clarity, and adds documentation and tests for the new filtering primitives.

* Fix topic group initialization in `SequenceTopicGroups::merge`

Corrects the initialization of SequenceTopicGroup by passing the accumulated topics instead of an empty vector when adding to the result.

mosaico-labs#12

* Add message after db and storage deletion in cleanup

After bringing down the docker compose environment in the cleanup function, a message is now displayed indicating that the database and storage have been deleted.

* Remove single ontology tag restriction in query builder

Eliminated the restriction that only allowed querying a single ontology tag per query in QueryOntologyCatalog. Updated integration tests to verify multi-tag queries are now supported.

Closes mosaico-labs#12

* (After manual merge) Refactor chunk filtering to use has_rows instead of count

Adapting the new and more performant logic to the `FacadeQuery`